# Conflicts:
#	Assets/Polytope Studio/Lowpoly_Demos/Environment_Free/Helpers/Plane mat.mat
#	Assets/Scripts/LMS/Card/Skill/LionRoar.cs
#	Assets/Scripts/Monster.cs

+++ Assets/Scripts/IDamageable.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 interface IDamageable
 {
-    void TakeDamage(int damage,Vector3 reactVec);
+    void TakeDamage(float damage,Vector3 reactVec);
 
-    int GetDamage();
+    float GetDamage();
 }
